Affichage modeles
o Chargement de modèles avec assimp
o Affichage ESGIGL + Repères by Augustin le fifou
o Shader diffuse - ambient - specular

+++ Projects/Model_OBJ/Mesh.hpp
@@ -10,11 +10,11 @@ class Mesh
 {
 	std::vector<vertexData> data;
 	std::vector<textureData> textures;
-	std::vector<unsigned int> indices;
+	std::vector<unsigned short> indices;
 	unsigned int vbo;
 	unsigned int ind;
 public:
-	Mesh(std::vector<vertexData> * vd, std::vector<unsigned int> * id, std::vector<textureData> * td = NULL);
+	Mesh(std::vector<vertexData> * vd, std::vector<unsigned short> * id, std::vector<textureData> * td = NULL);
 	~Mesh();
 	void draw(unsigned int programId);
 };
